refactoring runmat-parser for maintainability
