+++ crates/runmat-parser/src/span.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
+pub struct Span {
+    pub start: usize,
+    pub end: usize,
+}
